third_party/blink/web_tests/http/tests/plugins/cross-frame-object-access-expected.txt

CONSOLE ERROR: Uncaught SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
CONSOLE ERROR: Uncaught SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.


--------
Frame: 'childFrame'
--------
This tests that plugins can access objects in other frames as allowed by the security model enforced in WebCore.
Error: Error: Failed conversion between PP_Var and V8 value
Error: Uncaught
Error: Error: Failed conversion between PP_Var and V8 value
Error: Uncaught
SUCCESS
